#16066: Add seed param to uniform and bernoulli ops (#16179)
### Ticket
Link to Github Issue: #16066 

### Problem description
The uniform and bernoulli operation unit-tests behavior are
un-consistent because a random seed is created inside the program
factory.

### What's changed

- Add seed param to uniform and bernoulli operation, which allows
passing a seed to these ops in python side. If seed = 0 or not passed, a
random seed is generated in the program factory.
- Add a custom compute program hash function for these ops so that seed
param is not included in the the program hash.
- Update pytest for these operations

### Checklist
- [x] Post commit CI passes
https://github.com/tenstorrent/tt-metal/actions/runs/12411880762
- [ ] Blackhole Post commit (if applicable)
- [ ] Model regression CI testing passes (if applicable)
- [ ] Device performance regression CI testing passes (if applicable)
- [ ] **(For models and ops writers)** Full [new
models](https://github.com/tenstorrent/tt-metal/actions/workflows/full-new-models-suite.yaml)
tests passes
- [x] New/Existing tests provide coverage for changes

---------

Co-authored-by: Michael Chiou <156848643+ttmchiou@users.noreply.github.com>