Speed up linking for incremental builds.  (#15994)
### Problem description
Incremental builds take a very long time due to the time taken to link
`_ttnn.so`.

### What's changed
Split ttnn into multiple sub libraries, were each sub library is built
and linked into shared libraries independently. These sub libraries are
then linked together to form `_ttnn.so`

Reduced incremental build time from **150s** to **21s** for Debug builds
.

This speed up was achieved by using shared libraries for all the sub
libraries in Debug build. For Release build, there is only one libraries
so that there are not a lot of installation artifacts.

### New Flag for build_metal.sh
There is a new flag added to the build script called
`--ttnn-shared-sub-libs` that can be used to enable this feature. It is
off by default. It is best used to speed up Debug builds.

### Checklist
- [x] Post commit CI
[passes](https://github.com/tenstorrent/tt-metal/actions/runs/12351388706)