Extract Device interface (#16482)
### Ticket
None

### Problem description
Propagation of TT-Distributed entities like MeshDevice into TT-NN is
currently blocked.
TT-NN heavily relies on Device class and there is currently no way to
abstract MeshDevice behind it.
Providing such an abstraction is a critical next step to unblock
adoption of tt-distributed apis.

Related gh discussion
https://github.com/tenstorrent/tt-metal/discussions/16500

### What's changed
In previous PRs we moved Device members and some methods to the private
section, removed a couple friends.

The main change in this PR is an extraction an the Device interface.
This is the first step before the unification of Device / MeshDevice.
The PR contains multiple changes:
* Introduces IDevice (located in include/device.hpp, existing header
moved to deprecated)
* Removes all instances of "friend Device"
* Replaces relevant usage of Device with IDevice
* Removes DeviceHandle, replacing it with IDevice*
* Moves most binding to the base class, while still binding Device for
creation

#### Considerations:
* Existing code like this `Device* device = tensor.device();` in
3rdparty clients will not compile anymore and have to be changed to
`IDevice* device = ..` or `auto device = ...`. There is no remediation
to this.
* IDevice is not clean. We considered to first clean up the device by
moving ~50-60 methods out of it. Instead, we want to first extract the
interface to start unification with MeshDevice. The interface
improvement is separate from unblocking propagation of tt-distributed
entities into TT-NN
* Considered to re-split PR into pieces. Might possibly have to go this
route. Retrospectively it is clear that we can extract next pieces out:
  * Remove Device from friends
* Remove DeviceHandle or replace it with `using DeviceHandle = IDevice*`
  * Extract interface
  * Replace impl header include with interface include

#### Next steps:
Immediate next next step might be to inherit MeshDevice from IDevice, to
evolve the interface and to satisfy its needs.
There is a chance that as we work on that - we will have to change the
path and to switch to having two separate abstractions for a single
device (mock device, real pcie device) and a cluster. But even if we
turn, this change still makes sense and reduces leakage of details to
consumers as instead of including the impl header consumers now include
the interface.

### Checklist
- [x] [Post commit
CI](https://github.com/tenstorrent/tt-metal/actions/runs/12665570602)
- [x] [T3K
perf](https://github.com/tenstorrent/tt-metal/actions/runs/12664374781)
- [x] [Single Device Model
Perf](https://github.com/tenstorrent/tt-metal/actions/runs/12664388663)
- [x] [Single Device
Perf](https://github.com/tenstorrent/tt-metal/actions/runs/12666043923)
- [x] [Nightly model and ttnn
tests](https://github.com/tenstorrent/tt-metal/actions/runs/12666055919)

---------

Co-authored-by: Stanislav Minakov <sminakov@tenstorrent.com>