Add option to raise error on failed local/global tensor comparison
Enabling 'comparison_model_should_raise_exception' will cause the
program to crash when an invalid comparison is encountered. This is
useful for finding bugs in scenarios where a large number of operations
are being run.

This commit also adds a warning to tell users that comparison mode
cannot be used with fast dispatch.

Added documentation on comparison mode.