#16621: Add barriers at end of cq_dispatch_slave.cpp on IERISC
Like other kernels that can run on erisc, cq_dispatch_slave needs a barrier at
the end to ensure all transactions have finished. Otherwise, an assert will be
hit if the watcher is running.

 These barriers don't work on worker cores, because there cq_dispatch is on the
same core and shares use of this noc, but doesn't update this risc's
transaction counts. However, we don't have the barrier checks in brisck.cc, so
we can skip this for now.