#0: API Unification for Device and MeshDevice (#16570)
### Ticket
None

### Problem description
The motivation is to provide API unification between Device and
MeshDevice as we push virtualization over devices in a mesh down to
tt-metal. This is pre-requisite work needed for TT-Distributed.

At a high-level, the changes in this PR provides API unification between
Device/MeshDevice by simply implementing all required methods captured
in IDevice. This means APIs interfacing with an IDevice* can do so via a
MeshDevice through runtime polymorphism.

For APIs that can't sensibly be called through MeshDevice, we simply
throw an exception (for now). Unfortunately because we have such a large
surface area of APIs exposed through the IDevice interface, we are
currently required to implement all of these methods. However, there is
a separate effort to help minimize this surface area which is out of
scope for this PR.

### What's changed

- Explicitly make MeshDevice use the interface exposed by IDevice
- Move inf/nan/eps methods out of IDevice and into HAL

### Checklist

- [x] Post commit CI passes:
https://github.com/tenstorrent/tt-metal/actions/runs/12776971105
- [x] T3000 Regresion Tests:
https://github.com/tenstorrent/tt-metal/actions/runs/12776968242