Add tiled interleaved permute for when width dimension doesn't move (row-major tiled invariant) (#16671)
### Ticket
#16464
#16465

### Problem description
Currently tiled permute only works when the tiles are not changed at
all, and we attempt to convert it to row major before permuting
otherwise.

### What's changed
Add native support for tiled permute when the height dimension in the
tile is swapped elsewhere.

PyTorch2 transpose sweeps coverage increases to 100% with the following
caveats:
- for 3 large 2D tensors, the PCC drops from 0.999 to 0.998, which needs
to be investigated (this is in another kernel)
- for 1 large 2D tensor, the PCC check takes so long that it causes a
sweep timeout occasionally
Permute PyTorch2 transpose sweeps coverage increases from 94.5 to 95.4%
with this change. The PR for #16466 will increase this to 100%.
As part of this change, we also remove some more recursive transpose
chains in permute, increasing performance. For example:

| Input size     | Perm       | Percent speed-up |
|----------------|------------|------------------|
| 96,96,96,96    | 1,2,0,3    | 3.298783825      |
| 96,96,96,96    | 2,0,1,3    | 11.49442464      |
| 96,96,96,96    | 2,1,0,3    | 24.72870581      |


### Checklist
- [ ] Post commit CI passes
https://github.com/tenstorrent/tt-metal/actions/runs/12753938967
- [ ] Blackhole Post commit (if applicable)
https://github.com/tenstorrent/tt-metal/actions/runs/12750563591
- [ ] Model regression CI testing passes (if applicable)
https://github.com/tenstorrent/tt-metal/actions/runs/12750576671
- [ ] Device performance regression CI testing passes (if applicable)
https://github.com/tenstorrent/tt-metal/actions/runs/12718706607
- [ ] **(For models and ops writers)** Full [new
models](https://github.com/tenstorrent/tt-metal/actions/workflows/full-new-models-suite.yaml)
tests passes
- [ ] New/Existing tests provide coverage for changes