Fix test_new_all_gather.py regressions caused by API unification between Device/MeshDevice (#16836)
### Ticket
Link to Github Issue

### Problem description
After Device / MeshDevice API unification merge, one of the programs is
getting a core specified that is not one of its active eth cores. I
triaged and reverted the bad changes that cause this regression.

### What's changed
The change that I made seemed to have altered the device ordering
contained within MeshDevice. These new CCL tests seem to be sensitive to
the ordering of devices.

### Checklist
- [x] Post commit CI passes:
https://github.com/tenstorrent/tt-metal/actions/runs/12819606844
- [x] T3000 Tests:
https://github.com/tenstorrent/tt-metal/actions/runs/12819611934