#0: Split device_command_sequence in two
Currently we have a single device_command_sequence containing all the commands
for a program, but internally it's split into two parts between the CB writes
and the program binary writes. It would make more sense to have two separate
sequences, since that generalizes better and is simpler to use.