#0: Remove templating from process_write_linear.
The only real difference between the implementations is the call to
cq_noc_async_write_init_state outside the loop, so we can just combine the
implementations and choose which version of cq_noc_async_write_init_state to
call inside the implementation. This saves binary space for this kernel.