Restore clang-tidy scan to be incremental (#17073)
### Ticket
None

### Problem description
When I fixed the warnings-as-errors I mistakenly only updated it for the
scan. CMake (or Ninja) correctly noticed that the build command is
different and thus all targets are invalidated. That results in a full
scan instead of incremental.

### What's changed
Updated the baseline cmake configure to be the ultimately desired
command.
Deleted the scan cmake configure to avoid having them diverge again in
the future.

Deleted an old debug line that is no longer relevant now that we've
nuked LFS.