#16948: Add fixes to support unpadded inner dim for pf + mm and add support for DRAM in1 inputs for ring matmul. (#17311)
### Ticket
- #17060
- #16948

### Problem description
With the new validation being added to tensor_spec, the current
implementation of ring matmul with unpadded shapes fails.

Also, the matmul currently does not support DRAM interleaved in1
weights, which is required for matmuls with large weights that cannot
fit in L1 (such as the LM head in Llama).

### What's changed
- Internally round up the inner dim
- Add support for DRAM_INTERLEAVED in1