#0: Fix failing Llama TG tests by preserving old behavior for ShardTensorToMesh
Previously, when we had a MxN MeshDevice, a mesh_mapper of
ShardTensorToMesh would behave differently based on whether `mesh_type`
passed into the MeshDevice was MeshType::RowMajor, MeshType::Ring.

With the removal of `MeshType` from MeshDevice specification, this
changed the default behavior for users constructing a MeshDevice
with default mesh_type=MeshType::RowMajor. This change now preserves the
old behavior so that shards are distributed in row-major instead of a
line.