Remove non tile multiple width/height from conv2d
These two features are non critical for conv2d
meaning they don't contribute to enabling any model
perf on any model or improve pass rate on any sweep.

Problem with these features is that they kick in
in very unpredictable conditions for both users and
developers as they have many limits/conditions.

They are adding to conv2d test matrix, but they are
hard to test for as deriving tests that will trigger
them on multiple hw platforms is not easy.

Moreover they are source of bugs like #17647, and it's
often non obvious that bugs originate from these features
and when faced with a bug in conv2d first thing is to go to the
code and manually disable them to check for that.

For the reasons above these will get removed, and by
removing them #17647 will be fixed.