Cutover to the builders in CI v2 (#20284)
### Ticket
Somewhere

### Problem description
Follow-up to #20271 that doe the actual cutover. Revert this PR if
anything goes sideways.

### What's changed
* Switch to using the builders in CI v2
* Use Redis for the shared ccache
* Disable the local ccache (builders are ephemeral, so writing to disk
is pointless)
* Treat Redis as read-only when on branches
* This is enforced in the backend, so don't waste cycles trying to write
what cannot be written

Note: Not marking [skip ci] in order to have builds trigger and start
warming up Redis. Also in case anything goes sideways this PR is flagged
instead of the innocent follower.