tt_transformers: Enable support for non-uniform decoder configurations (#19431)
**Problem description**
Currently, all the decoders in a model have the same computation
precision.
Ideally we would want to configure each layer individually to explore
possible optimizations.

**What's changed**
This PR allows us to provide a json file (or a folder with several json
files), e.g.,
```
"decoders": {
        "0": {
            "precision_cfg": { "FF1_FF3": "BFP8", "FF2": "BFP8" },
            "fidelity_cfg": { }
        },
        "1": {
            "precision_cfg": { "FF1_FF3": "BFP8", "FF2": "BFP4", "KV_CACHE": "BF16" },
            "fidelity_cfg": { "LI_FF1_FF3": "HIFI2" }
        },
        "2": {
            "precision_cfg": { "FF1_FF3": "BF16", "WQKV": "BF16", "WO": "BFP8", "KV_CACHE": "BFP8"},
            "fidelity_cfg": { "SDPA_DECODE": "HIFI2_NA", "LI_QKV_DECODE": "HIFI4" }
        },
...
```
which sets the configuration of each decoder. When a component is not
provided (i.e. FF2 is missing for decoder 2), the default configuration
for that component is the one that is set.
We can also run the `lt` tool for the configurations in the folder, and
create the Pareto graph for them:
<img width="673" alt="Screenshot 2025-04-03 at 6 18 37 PM"
src="https://github.com/user-attachments/assets/41d728f8-d8c5-4459-9a1f-38c6599f975d"
/>
Note that the configuration description in the graph is removed when the
configuration is read from a json file, since the entire description is
too long. In addition, when the configuration is read from a json file,
the result is not compared against PERF.md, and is also not written no
PERF.md

---------

Signed-off-by: Tomer Levin <tlevin@tenstorrent.com>