[skip ci] Run build + smoke of MergeGate on main (#20527)
### Ticket
N/A

### Problem description
We're skipping the build when on main (b/c no delta from 'main'). But we
need to build on main to populate ccache.
While we're there, we may as well also run the Smoke which is
lightweight and provides good evidence that Main remains clean and so
any failure is purely from the branch.

### What's changed
Adjusted conditionals to run builds + smoke when triggered on main.