#20867: Modify test_pgm_dispatch to test with multiple subdevice core ranges
Subdevices that need multiple core ranges to specify cores can be slower to
run, since multiple GO messages need to be sent. Add a test for that.

Incidentally, this also fixes a bug in the KG tests where all KGs except the
first only consisted of a single core (instead of an entire column).